added Settingsd to HamburgerMenu #13
