+++ Drachenhorn.Desktop/Views/MainView.xaml
@@ -11,6 +11,7 @@
                  xmlns:helper="clr-namespace:Drachenhorn.Desktop.Helper"
                  xmlns:common="clr-namespace:Drachenhorn.Core.ViewModels.Common;assembly=Drachenhorn.Core"
                  xmlns:template="clr-namespace:Drachenhorn.Desktop.UserControls.Template"
+                 xmlns:common1="clr-namespace:Drachenhorn.Desktop.UserControls.Common"
                  x:Class="Drachenhorn.Desktop.Views.MainView" mc:Ignorable="d"
                  WindowState="Maximized" TitleCharacterCasing="Normal" ShowIconOnTitleBar="True"
                  ShowTitleBar="False" Closing="MainView_OnClosing" Loaded="MainView_OnLoaded"
@@ -55,7 +56,8 @@
                            SelectedIndex="0"
                            DisplayMode="CompactInline"
                            Name="HamburgerMenuControl"
-                           ItemClick="HamburgerMenu_OnItemClick">
+                           ItemClick="HamburgerMenu_OnItemClick"
+                           OptionsItemClick="HamburgerMenu_OnItemClick">
             <mah:HamburgerMenu.ContentTemplate>
                 <DataTemplate DataType="{x:Type mah:HamburgerMenuItem}">
                     <Grid x:Name="TheContentGrid">
@@ -192,8 +194,11 @@
             <mah:HamburgerMenu.OptionsItemsSource>
                 <mah:HamburgerMenuItemCollection>
                     <mah:HamburgerMenuIconItem Icon="{StaticResource PropertiesIconWhite}"
-                                               Label="{lang:Translate UI.Properties}"
-                                               Command="{Binding ShowSettings}"/>
+                                               Label="{lang:Translate UI.Properties}">
+                        <mah:HamburgerMenuItem.Tag>
+                            <common1:SettingsViewControl/>
+                        </mah:HamburgerMenuItem.Tag>
+                    </mah:HamburgerMenuIconItem>
                 </mah:HamburgerMenuItemCollection>
             </mah:HamburgerMenu.OptionsItemsSource>
         </mah:HamburgerMenu>
