+++ Drachenhorn.Desktop/Views/SettingsView.xaml.cs
@@ -20,51 +20,6 @@ namespace Drachenhorn.Desktop.Views
         public SettingsView()
         {
             InitializeComponent();
-
-            Loaded += (sender, args) => { LanguageComboBox_SelectionChanged(sender, null); };
-        }
-
-        private void Hyperlink_RequestNavigate(object sender, RequestNavigateEventArgs e)
-        {
-            Process.Start(new ProcessStartInfo(e.Uri.AbsoluteUri));
-            e.Handled = true;
-        }
-
-        private void LanguageComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            if (FlagImage == null) return;
-
-            try
-            {
-                var culture = SimpleIoc.Default.GetInstance<ISettings>().CurrentCulture.Name;
-
-                var path = "pack://application:,,,/Images/Flags/" + culture + ".png";
-                FlagImage.Source = new BitmapImage(new Uri(path));
-            }
-            catch (IOException)
-            {
-                var path = "pack://application:,,,/Images/Flags/invariant.png";
-                FlagImage.Source = new BitmapImage(new Uri(path));
-            }
-        }
-
-        private static BitmapImage LoadImage(byte[] imageData)
-        {
-            if (imageData == null || imageData.Length == 0) return null;
-            var image = new BitmapImage();
-            using (var mem = new MemoryStream(imageData))
-            {
-                mem.Position = 0;
-                image.BeginInit();
-                image.CreateOptions = BitmapCreateOptions.PreservePixelFormat;
-                image.CacheOption = BitmapCacheOption.OnLoad;
-                image.UriSource = null;
-                image.StreamSource = mem;
-                image.EndInit();
-            }
-
-            image.Freeze();
-            return image;
         }
     }
 }
